nx_kit: Minor warning fixes
<==== Mercurial data ====>
Branch: meta_4.0

+++ artifacts/nx_kit/src/nx/kit/test.cpp
@@ -275,7 +275,7 @@ static void processArgs(int argc, const char* const argv[])
         return; //< Do nothing if no args were specified.
 
     const std::vector<std::string> args{argv, argv + argc};
-    const auto arg = [&args](int i) { return (args.size() <= i) ? "" : args[i]; };
+    const auto arg = [&args](int i) { return ((int) args.size() <= i) ? "" : args[i]; };
 
     if (args.size() == 2 && (arg(1) == "-h" || arg(1) == "--help"))
     {
@@ -365,7 +365,7 @@ int runAllTests(const char* testSuiteName, int argc, const char* const argv[])
         << "Running " << allTests().size() << " test(s) from " << fullSuiteName << std::endl;
 
     std::vector<int> failedTests;
-    for (int i = 1; i <= allTests().size(); ++i)
+    for (int i = 1; i <= (int) allTests().size(); ++i)
     {
         if (!runTest(allTests()[i - 1], i))
             failedTests.push_back(i);
